Moved content of future_1_5 in the new compat file
