+++ django_extensions/management/commands/print_user_for_session.py
@@ -3,11 +3,7 @@ from django.core.management.base import BaseCommand, CommandError
 
 from django_extensions.compat import importlib
 from django_extensions.management.utils import signalcommand
-
-try:
-    from django.contrib.auth import get_user_model  # Django 1.5
-except ImportError:
-    from django_extensions.future_1_5 import get_user_model
+from django_extensions.compat import get_user_model
 
 try:
     from django.contrib.sessions.backends.base import VALID_KEY_CHARS  # Django 1.5
