+++ django_extensions/management/commands/set_fake_emails.py
@@ -41,10 +41,7 @@ class Command(NoArgsCommand):
         if not settings.DEBUG:
             raise CommandError('Only available in debug mode')
 
-        try:
-            from django.contrib.auth import get_user_model  # Django 1.5
-        except ImportError:
-            from django_extensions.future_1_5 import get_user_model
+        from django_extensions.compat import get_user_model
         from django.contrib.auth.models import Group
         email = options.get('default_email', DEFAULT_FAKE_EMAIL)
         include_regexp = options.get('include_regexp', None)
